backendd/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
asyncpg==0.30.0
alembic==1.15.2
python-dotenv==1.1.0